+++ experimental/traffic-portal/src/app/core/cache-groups/cache-group-details/cache-group-details.component.ts
@@ -12,15 +12,14 @@
 * limitations under the License.
 */
 
-import { Location } from "@angular/common";
-import { Component, type OnInit } from "@angular/core";
+import { Component, OnInit } from "@angular/core";
 import { FormControl } from "@angular/forms";
 import { MatDialog } from "@angular/material/dialog";
-import { ActivatedRoute } from "@angular/router";
-import { LocalizationMethod, localizationMethodToString, TypeFromResponse, type ResponseCacheGroup } from "trafficops-types";
+import { ActivatedRoute, Router } from "@angular/router";
+import { LocalizationMethod, localizationMethodToString, TypeFromResponse, ResponseCacheGroup } from "trafficops-types";
 
 import { CacheGroupService, TypeService } from "src/app/api";
-import { DecisionDialogComponent, type DecisionDialogData } from "src/app/shared/dialogs/decision-dialog/decision-dialog.component";
+import { DecisionDialogComponent, DecisionDialogData } from "src/app/shared/dialogs/decision-dialog/decision-dialog.component";
 import { LoggingService } from "src/app/shared/logging.service";
 import { NavigationService } from "src/app/shared/navigation/navigation.service";
 
@@ -69,9 +68,9 @@ export class CacheGroupDetailsComponent implements OnInit {
 
 	constructor(
 		private readonly route: ActivatedRoute,
+		private readonly router: Router,
 		private readonly api: CacheGroupService,
 		private readonly typesAPI: TypeService,
-		private readonly location: Location,
 		private readonly dialog: MatDialog,
 		private readonly navSvc: NavigationService,
 		private readonly log: LoggingService,
@@ -182,7 +181,7 @@ export class CacheGroupDetailsComponent implements OnInit {
 		ref.afterClosed().subscribe(result => {
 			if (result) {
 				this.api.deleteCacheGroup(this.cacheGroup);
-				this.location.replaceState("core/cache-groups");
+				this.router.navigate(["core/cache-groups"]);
 			}
 		});
 	}
@@ -208,6 +207,7 @@ export class CacheGroupDetailsComponent implements OnInit {
 		if (this.new) {
 			this.cacheGroup = await this.api.createCacheGroup(this.cacheGroup);
 			this.new = false;
+			await this.router.navigate(["core/cache-groups", this.cacheGroup.id]);
 		} else {
 			this.cacheGroup = await this.api.updateCacheGroup(this.cacheGroup);
 		}