+++ experimental/traffic-portal/src/app/core/parameters/detail/parameter-detail.component.ts
@@ -12,10 +12,9 @@
 * limitations under the License.
 */
 
-import { Location } from "@angular/common";
 import { Component, OnInit } from "@angular/core";
 import { MatDialog } from "@angular/material/dialog";
-import { ActivatedRoute } from "@angular/router";
+import { ActivatedRoute, Router } from "@angular/router";
 import { ResponseParameter } from "trafficops-types";
 
 import { ProfileService } from "src/app/api";
@@ -41,8 +40,8 @@ export class ParameterDetailComponent implements OnInit {
 
 	constructor(
 		private readonly route: ActivatedRoute,
+		private readonly router: Router,
 		private readonly profileService: ProfileService,
-		private readonly location: Location,
 		private readonly dialog: MatDialog,
 		private readonly navSvc: NavigationService,
 		private readonly log: LoggingService,
@@ -58,8 +57,10 @@ export class ParameterDetailComponent implements OnInit {
 			return;
 		}
 
-		if (ID === "new") {
-			this.navSvc.headerTitle.next("New Parameter");
+		this.new = ID === "new";
+
+		if (this.new) {
+			this.setTitle();
 			this.new = true;
 			this.parameter = {
 				configFile: "",
@@ -80,7 +81,17 @@ export class ParameterDetailComponent implements OnInit {
 		}
 
 		this.parameter = await this.profileService.getParameters(numID);
-		this.navSvc.headerTitle.next(`Parameter: ${this.parameter.name} (${this.parameter.id})`);
+		this.setTitle();
+	}
+
+	/**
+	 * Sets the headerTitle based on current Parameter state.
+	 *
+	 * @private
+	 */
+	private setTitle(): void {
+		const title = this.new ? "New Parameter" : `Parameter: ${this.parameter.name} (${this.parameter.id})`;
+		this.navSvc.headerTitle.next(title);
 	}
 
 	/**
@@ -98,7 +109,7 @@ export class ParameterDetailComponent implements OnInit {
 		ref.afterClosed().subscribe(result => {
 			if(result) {
 				this.profileService.deleteParameter(this.parameter.id);
-				this.location.back();
+				this.router.navigate(["core/parameters"]);
 			}
 		});
 	}
@@ -114,8 +125,10 @@ export class ParameterDetailComponent implements OnInit {
 		if(this.new) {
 			this.parameter = await this.profileService.createParameter(this.parameter);
 			this.new = false;
+			await this.router.navigate(["core/parameters", this.parameter.id]);
 		} else {
 			this.parameter = await this.profileService.updateParameter(this.parameter);
 		}
+		this.setTitle();
 	}
 }