+++ experimental/traffic-portal/src/app/core/servers/capabilities/capabilities.component.ts
@@ -11,7 +11,8 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-import { Component, type OnInit } from "@angular/core";
+
+import { Component, OnInit } from "@angular/core";
 import { FormControl } from "@angular/forms";
 import { MatDialog } from "@angular/material/dialog";
 import { ActivatedRoute } from "@angular/router";