+++ experimental/traffic-portal/src/app/core/types/detail/type-detail.component.ts
@@ -12,10 +12,9 @@
 * limitations under the License.
 */
 
-import { Location } from "@angular/common";
 import { Component, OnInit } from "@angular/core";
 import { MatDialog } from "@angular/material/dialog";
-import { ActivatedRoute } from "@angular/router";
+import { ActivatedRoute, Router } from "@angular/router";
 import { TypeFromResponse } from "trafficops-types";
 
 import { TypeService } from "src/app/api";
@@ -37,8 +36,8 @@ export class TypeDetailComponent implements OnInit {
 
 	constructor(
 		private readonly route: ActivatedRoute,
+		private readonly router: Router,
 		private readonly typeService: TypeService,
-		private readonly location: Location,
 		private readonly dialog: MatDialog,
 		private readonly navSvc: NavigationService,
 		private readonly log: LoggingService,
@@ -54,8 +53,10 @@ export class TypeDetailComponent implements OnInit {
 			return;
 		}
 
-		if (ID === "new") {
-			this.navSvc.headerTitle.next("New Type");
+		this.new = ID === "new";
+
+		if (this.new) {
+			this.setTitle();
 			this.new = true;
 			this.type = {
 				description: "",
@@ -74,7 +75,17 @@ export class TypeDetailComponent implements OnInit {
 		}
 
 		this.type = await this.typeService.getTypes(numID);
-		this.navSvc.headerTitle.next(`Type: ${this.type.name}`);
+		this.setTitle();
+	}
+
+	/**
+	 * Sets the headerTitle based on current Type state.
+	 *
+	 * @private
+	 */
+	private setTitle(): void {
+		const title = this.new ? "New Type" : `Type: ${this.type.name}`;
+		this.navSvc.headerTitle.next(title);
 	}
 
 	/**
@@ -92,7 +103,7 @@ export class TypeDetailComponent implements OnInit {
 		ref.afterClosed().subscribe(result => {
 			if(result) {
 				this.typeService.deleteType(this.type.id);
-				this.location.back();
+				this.router.navigate(["core/types"]);
 			}
 		});
 	}
@@ -108,9 +119,11 @@ export class TypeDetailComponent implements OnInit {
 		if(this.new) {
 			this.type = await this.typeService.createType(this.type);
 			this.new = false;
+			await this.router.navigate(["core/types", this.type.id]);
 		} else {
 			this.type = await this.typeService.updateType(this.type);
 		}
+		this.setTitle();
 	}
 
 }